refactor - favorites: 패키지 명 변경
